Move smsService class to the Service directory
